code refactored, int json objects changed from float

+++ app/src/main/java/one/mann/weatherman/model/openWeatherMap/Clouds.kt
@@ -1,5 +1,5 @@
 package one.mann.weatherman.model.openWeatherMap
 
 class Clouds {
-    val all: Int = 0
+    val all: Float = 0.toFloat()
 }
