+++ app/src/main/java/one/mann/weatherman/model/openWeatherMap/Wind.kt
@@ -2,5 +2,5 @@ package one.mann.weatherman.model.openWeatherMap
 
 class Wind {
     val speed: Float = 0.toFloat()
-    val deg: Int = 0
+    val deg: Float = 0.toFloat()
 }
